Patch submitted by Ludovic Courtès:
Fix `jde-help-find-javadoc' so that it correctly infers whether
an HTML page exists.

Previously it would consider that Wget succeeded if its output
contained the string "200".  This doesn't work as Wget's output
starts with something like:

  --2009-07-10 16:15:37--  http://java.sun.com/javase/6/docs/api/foobar

This patch makes it check the exit code of Wget instead, which is
more reliable.



git-svn-id: https://svn.code.sf.net/p/jdee/code/branches/2.4.0/jde@92 381a1f10-3442-0410-997e-ef331cd04102