Add name slot that's no longer in EIEIO - fixes #53
"The problem is that in the latest version of eieio instantiating a class with a
string arg no longer generates an automatic name member. I have fixed this
problem in the class jdee-url by adding an explict name member and modifying the
few places the name was referenced and where instantiated."

Contributed by fseibel.