Test files have different target directories than program files when compiled
When compiling a test file its target directory is normally something bellow "test-classes". Regular classes are compiled to "classes" directory. Thus regular classes and test classes have different jdee-compile-option-directory in the same project.
A solution to this issue is to make jdee-compile-option-directory a buffer local variable and set its value accordingly in prj.el . But flycheck cannot known where to put the compiled file, so the solution I propose is to copy jdee-compile-option-directory value from original buffer if it was set.