Activate main method abbreviation
According to the documentation `(info "(jdee) Control Flow
Abbreviations")`, there should be a main method abbreviation bound to
"main".  However, abbreviations like "ife", "try" work fine in JDEE,
but "main" doesn't.  This was because the `abbrev` argument passed to
`tempo-define-template` was nil.

Rewrite the `:set` part using `jdee-gen-define-abbrev-template` and
move the `defcustom` after the definition of that function to avoid a
`void-function` error.