Update .dir-locals.el
Simplify `.dir-locals.el` to make it more human readable.  And add
`(lisp-indent-function . lisp-indent-function)` to it.  However, you
may also need to add this to your `safe-local-variable-values`.  This
is definitely safe, since this is actually the default value for
Emacs.  Many Lisp programmers like to set the indent function to
something else like `common-lisp-indent-function`.  This way, they
don't have to change this variable back and forth when switching
projects and it's likely for this project to get a consistent
indentation.